move add demo logic into its own folder
